style(esl-utils): `misc/object` split between the files
