+++ src/modules/esl-utils/misc/object/compare.ts
@@ -1,0 +1,14 @@
+import {isObject} from './types';
+
+/** Deep object compare */
+export function deepCompare(obj1: any, obj2: any): boolean {
+  if (Object.is(obj1, obj2)) return true;
+  if (typeof obj1 !== typeof obj2) return false;
+  if (isObject(obj1) && isObject(obj2)) {
+    const keys1 = Object.keys(obj1);
+    const keys2 = Object.keys(obj2);
+    if (keys1.length !== keys2.length) return false;
+    return !keys1.some((key) => !deepCompare(obj1[key], obj2[key]));
+  }
+  return false;
+}
